% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_flood_days.R
\name{count_flood_days}
\alias{count_flood_days}
\title{Count flood days from river height data}
\usage{
count_flood_days(river_height_data, flood_threshold = 20)
}
\arguments{
\item{river_height_data}{A data frame with a river_ht containing daily river heights.}

\item{flood_threshold}{River height above which a day counts as a flood. Default is 20.}
}
\value{
Numeric: The number of days where river height exceeded the threshold.
}
\description{
Counts the number of days where river height exceeds a flood threshold
}
