% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatal_cost.R
\name{fatal_cost}
\alias{fatal_cost}
\title{Cost of Loss of Human Life}
\usage{
fatal_cost(crashcount, vsl, discount, time)
}
\arguments{
\item{crashcount}{number of fatalities per year due to flooding (count)}

\item{vsl}{value of a statistical life (USD) default = 7400000}

\item{discount}{discount rate (no unit) default = 7\%}

\item{time}{number of years evaluation default = 30 years}
}
\value{
npv (USD) net present value
}
\description{
This function determines the total cost of human life from bridge overtopping
}
